[18110] Fixed type cast bug.
git-svn-id: https://svn.dbc.dk/repos/dataio/trunk@80305 b87a6194-a221-0410-ad6f-3b8ec35bba13

+++ integration-test/new-job-store-service/src/test/java/dk/dbc/dataio/jobstore/service/ejb/PgJobStoreIT.java
@@ -197,7 +197,7 @@ public class PgJobStoreIT {
         // Given...
         final PgJobStore pgJobStore = newPgJobStore();
         final ChunkEntity chunk = new ChunkEntity();
-        chunk.setNumberOfItems(4);
+        chunk.setNumberOfItems((short) 4);
         chunk.setDataFileId("datafileID");
         chunk.setSequenceAnalysisData(new SequenceAnalysisData(new HashSet(Arrays.asList("sequence analysis data"))));
         chunk.setState(new State());
